backend/requirements.txt
flask==2.2.3
werkzeug==2.2.3
flask-cors==3.0.10
flask-jwt-extended==4.4.4
pydantic==1.10.7
bcrypt==4.0.1
python-dotenv==1.0.0 